requirements.txt
herepy
python-dotenv
daiquiri
